feat: make global page the homepage, move pools to /pools
- / → Global overview (was /global)
- /pools → Pool list + swap feed (was /)
- Nav: Global | Pools
- Pool detail breadcrumb: Home → Global
