--- ui-dashboard/src/app/global/page.tsx
@@ -1,180 +1,0 @@
-"use client";
-
-import { Suspense } from "react";
-import Link from "next/link";
-import { useGQL } from "@/lib/graphql";
-import { ALL_POOLS_WITH_HEALTH } from "@/lib/queries";
-import { formatWei } from "@/lib/format";
-import { poolName, isFpmm } from "@/lib/tokens";
-import { useNetwork } from "@/components/network-provider";
-import type { Pool } from "@/lib/types";
-import { Table, Row, Th, Td } from "@/components/table";
-import { Skeleton, EmptyBox, ErrorBox, Tile } from "@/components/feedback";
-import { SourceBadge } from "@/components/badges";
-import { PoolsTable } from "@/components/pools-table";
-
-export default function GlobalPage() {
-  return (
-    <Suspense>
-      <GlobalContent />
-    </Suspense>
-  );
-}
-
-function GlobalContent() {
-  const {
-    data: poolsData,
-    error: poolsErr,
-    isLoading: poolsLoading,
-  } = useGQL<{ Pool: Pool[] }>(ALL_POOLS_WITH_HEALTH);
-
-  const pools = poolsData?.Pool ?? [];
-
-  const fpmmPools = pools.filter(isFpmm);
-  const virtualPools = pools.filter((p) => !isFpmm(p));
-
-  const okCount = pools.filter((p) => p.healthStatus === "OK").length;
-  const warnCount = pools.filter((p) => p.healthStatus === "WARN").length;
-  const critCount = pools.filter((p) => p.healthStatus === "CRITICAL").length;
-  // N/A = VirtualPools (oracle health not applicable) + any pools without health data
-  const naCount = pools.filter(
-    (p) => !p.healthStatus || p.healthStatus === "N/A",
-  ).length;
-
-  // Derive total swaps from pool cumulative counts (avoids a second query)
-  const totalSwaps = pools.reduce((sum, p) => sum + (p.swapCount ?? 0), 0);
-
-  return (
-    <div className="space-y-8">
-      <div>
-        <h1 className="text-2xl font-bold text-white mb-1">Global Overview</h1>
-        <p className="text-sm text-slate-400">
-          Protocol-wide statistics across all pools
-        </p>
-      </div>
-
-      {poolsErr && (
-        <ErrorBox message={`Failed to load data: ${poolsErr.message}`} />
-      )}
-
-      {/* Summary tiles */}
-      <section>
-        <h2 className="text-lg font-semibold text-white mb-3">Summary</h2>
-        <div className="grid grid-cols-2 sm:grid-cols-4 gap-4">
-          <Tile
-            label="Total Pools"
-            value={poolsLoading ? "…" : String(pools.length)}
-          />
-          <Tile
-            label="FPMM Pools"
-            value={poolsLoading ? "…" : String(fpmmPools.length)}
-          />
-          <Tile
-            label="Virtual Pools"
-            value={poolsLoading ? "…" : String(virtualPools.length)}
-          />
-          <Tile
-            label="Total Swaps"
-            value={poolsLoading ? "…" : totalSwaps.toLocaleString()}
-          />
-        </div>
-      </section>
-
-      {/* Health breakdown */}
-      <section>
-        <h2 className="text-lg font-semibold text-white mb-3">
-          Health Status
-          <span className="ml-2 text-xs font-normal text-slate-500">
-            (N/A = VirtualPools — oracle health not applicable)
-          </span>
-        </h2>
-        <div className="grid grid-cols-2 sm:grid-cols-4 gap-4">
-          <Tile label="🟢 OK" value={poolsLoading ? "…" : String(okCount)} />
-          <Tile label="🟡 WARN" value={poolsLoading ? "…" : String(warnCount)} />
-          <Tile
-            label="🔴 CRITICAL"
-            value={poolsLoading ? "…" : String(critCount)}
-          />
-          <Tile label="⚪ N/A" value={poolsLoading ? "…" : String(naCount)} />
-        </div>
-      </section>
-
-      {/* All pools table */}
-      <section>
-        <h2 className="text-lg font-semibold text-white mb-3">All Pools</h2>
-        {poolsLoading ? (
-          <Skeleton rows={5} />
-        ) : pools.length === 0 ? (
-          <EmptyBox message="No pools found." />
-        ) : (
-          <PoolsTable pools={pools} />
-        )}
-      </section>
-
-      {/* Activity summary */}
-      <section>
-        <h2 className="text-lg font-semibold text-white mb-3">
-          Pool Activity
-        </h2>
-        {poolsLoading ? (
-          <Skeleton rows={5} />
-        ) : pools.length === 0 ? (
-          <EmptyBox message="No pools found." />
-        ) : (
-          <ActivityTable pools={pools} />
-        )}
-      </section>
-    </div>
-  );
-}
-
-function ActivityTable({ pools }: { pools: Pool[] }) {
-  const { network } = useNetwork();
-  // Sort by swap count descending
-  const sorted = [...pools].sort(
-    (a, b) => (b.swapCount ?? 0) - (a.swapCount ?? 0),
-  );
-  return (
-    <Table>
-      <thead>
-        <tr className="border-b border-slate-800 bg-slate-900/50">
-          <Th>Pool</Th>
-          <Th>Type</Th>
-          <Th align="right">Swaps</Th>
-          <Th align="right">Rebalances</Th>
-          <Th align="right">Volume (Token 0)</Th>
-          <Th align="right">Volume (Token 1)</Th>
-        </tr>
-      </thead>
-      <tbody>
-        {sorted.map((p) => (
-          <Row key={p.id}>
-            <td className="px-4 py-3">
-              <Link
-                href={`/pool/${encodeURIComponent(p.id)}`}
-                className="font-semibold text-indigo-400 hover:text-indigo-300"
-              >
-                {poolName(network, p.token0, p.token1)}
-              </Link>
-            </td>
-            <td className="px-4 py-3">
-              <SourceBadge source={p.source} />
-            </td>
-            <Td mono small align="right">
-              {p.swapCount ?? 0}
-            </Td>
-            <Td mono small align="right">
-              {p.rebalanceCount ?? 0}
-            </Td>
-            <Td mono small align="right">
-              {p.notionalVolume0 ? formatWei(p.notionalVolume0) : "—"}
-            </Td>
-            <Td mono small align="right">
-              {p.notionalVolume1 ? formatWei(p.notionalVolume1) : "—"}
-            </Td>
-          </Row>
-        ))}
-      </tbody>
-    </Table>
-  );
-}
